Open folders as converation are opened
